Update README.md
I wanted to try something new with React, so I built this simple app. It's a small thing, but it helped me learn about how React works.

You can use the app to count up, count down, and reset the counter with just a few clicks. It might seem basic, but it's a step forward for me as a developer.